+++ src/main/java/net/frozenblock/lib/FrozenLibClient.java
@@ -21,6 +21,7 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientChunkEvents;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientWorldEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.frozenblock.lib.block.sound.impl.BlockSoundTypeManager;
@@ -37,6 +38,7 @@ import net.frozenblock.lib.registry.client.FrozenLibClientRegistries;
 import net.frozenblock.lib.screenshake.api.client.ScreenShaker;
 import net.frozenblock.lib.sound.client.impl.FlyBySoundHub;
 import net.frozenblock.lib.wind.client.impl.ClientWindManager;
+import net.frozenblock.lib.worldgen.structure.api.status.client.ClientStructureStatuses;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.protocol.common.ServerboundCustomPayloadPacket;
 import net.minecraft.resources.ResourceLocation;
@@ -82,15 +84,17 @@ public final class FrozenLibClient implements ClientModInitializer {
 			}
 		);
 		ClientTickEvents.START_CLIENT_TICK.register(client -> ClientWindManager.clearAndSwitchWindDisturbances());
-		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
-			ScreenShaker.clear();
-			ClientWindManager.reset();
+		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> clearClientListHolders());
+		ClientWorldEvents.AFTER_CLIENT_WORLD_CHANGE.register((minecraft, clientLevel) -> clearClientListHolders());
+		ClientChunkEvents.CHUNK_LOAD.register((world, chunk) -> {
+			if (FrozenLibConfig.IS_DEBUG) world.sendPacketToServer(new ServerboundCustomPayloadPacket(new StructureDebugRequestPayload(chunk.getPos())));
 		});
-		ClientChunkEvents.CHUNK_LOAD.register(
-			(world, chunk) -> {
-				if (FrozenLibConfig.IS_DEBUG) world.sendPacketToServer(new ServerboundCustomPayloadPacket(new StructureDebugRequestPayload(chunk.getPos())));
-			}
-		);
+	}
+
+	private static void clearClientListHolders() {
+		ScreenShaker.clear();
+		ClientWindManager.reset();
+		ClientStructureStatuses.clearStructureStatuses();
 	}
 
 }