+++ src/main/java/net/frozenblock/lib/networking/FrozenClientNetworking.java
@@ -70,6 +70,8 @@ import net.frozenblock.lib.wind.api.WindDisturbanceLogic;
 import net.frozenblock.lib.wind.client.impl.ClientWindManager;
 import net.frozenblock.lib.wind.impl.networking.WindDisturbancePacket;
 import net.frozenblock.lib.wind.impl.networking.WindSyncPacket;
+import net.frozenblock.lib.worldgen.structure.api.status.client.ClientStructureStatuses;
+import net.frozenblock.lib.worldgen.structure.impl.status.networking.PlayerStructureStatusPacket;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.multiplayer.ClientPacketListener;
@@ -114,6 +116,7 @@ public final class FrozenClientNetworking {
 		receiveIconRemovePacket();
 		receiveWindSyncPacket();
 		receiveWindDisturbancePacket();
+		receiveStructureStatusPacket();
 		receiveTransferImagePacket();
 		receiveCapePacket();
 		receiveCapeRepoPacket();
@@ -397,6 +400,13 @@ public final class FrozenClientNetworking {
 					)
 				);
 			}
+		});
+	}
+
+	@ApiStatus.Internal
+	private static void receiveStructureStatusPacket() {
+		ClientPlayNetworking.registerGlobalReceiver(PlayerStructureStatusPacket.PACKET_TYPE, (packet, ctx) -> {
+			ClientStructureStatuses.setStructureStatuses(packet.structureStatuses());
 		});
 	}