+++ CHANGELOG.md
@@ -5,4 +5,5 @@ Make sure to clear this after each release
 Put changelog here:
 
 -----------------
+- Added `StructureMusicApi,` allowing modders to define custom music to play within structures.
 - Fixed spam-logging of structure names upon world load.